+++ components/index.ts
@@ -5,3 +5,4 @@ export { default as Header } from './Header/Header';
 export { default as Card } from './Card/Card';
 export { default as Button } from './Button/Button';
 export { default as Icon } from './Icon/Icon';
+export { default as CustomSelectMenuWrapper } from './CustomSelectMenu/CustomSelectMenu';
